requirements.txt
numpy==1.20.2
opencv-python==4.5.2.52
Pillow==8.2.0
PyYAML==5.4.1
scipy==1.6.3
tqdm==4.60.0
